# docs/contracts/processing-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GNTV DIGITAL Processing Operations Center API",
    "version": "5.3",
    "description": "Standalone contract exported from the currently registered FastAPI processing routes."
  },
  "paths": {
    "/api/v1/processing/jobs": {
      "post": {
        "tags": [
          "Processing Operations Center"
        ],
        "summary": "Create Job",
        "operationId": "create_job_api_v1_processing_jobs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessingJobCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingJobResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user lacks the required processing permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The requested processing operation conflicts with current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Processing Operations Center"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_v1_processing_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingJobPage"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user lacks the required processing permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processing/jobs/{job_id}": {
      "get": {
        "tags": [
          "Processing Operations Center"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_processing_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingJobResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user lacks the required processing permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested processing resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Processing Operations Center"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_v1_processing_jobs__job_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingJobCancelResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user lacks the required processing permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested processing resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The requested processing operation conflicts with current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processing/workers": {
      "get": {
        "tags": [
          "Processing Operations Center"
        ],
        "summary": "Workers",
        "operationId": "workers_api_v1_processing_workers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkerStatusResponse"
                  },
                  "type": "array",
                  "title": "Response Workers Api V1 Processing Workers Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user lacks the required processing permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/processing/queues": {
      "get": {
        "tags": [
          "Processing Operations Center"
        ],
        "summary": "Queues",
        "operationId": "queues_api_v1_processing_queues_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTelemetryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user lacks the required processing permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/processing/manifests": {
      "get": {
        "tags": [
          "Processing Operations Center"
        ],
        "summary": "Manifests",
        "operationId": "manifests_api_v1_processing_manifests_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManifestPage"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user lacks the required processing permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processing/manifests/validate": {
      "post": {
        "tags": [
          "Processing Operations Center"
        ],
        "summary": "Validate Manifest",
        "operationId": "validate_manifest_api_v1_processing_manifests_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManifestValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManifestValidationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user lacks the required processing permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/processing/manifests/{manifest_id}": {
      "get": {
        "tags": [
          "Processing Operations Center"
        ],
        "summary": "Manifest",
        "operationId": "manifest_api_v1_processing_manifests__manifest_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "manifest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Manifest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManifestStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user lacks the required processing permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested processing resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processing/thumbnails": {
      "get": {
        "tags": [
          "Processing Operations Center"
        ],
        "summary": "Thumbnails",
        "operationId": "thumbnails_api_v1_processing_thumbnails_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThumbnailPage"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user lacks the required processing permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processing/thumbnails/{thumbnail_id}": {
      "get": {
        "tags": [
          "Processing Operations Center"
        ],
        "summary": "Thumbnail",
        "operationId": "thumbnail_api_v1_processing_thumbnails__thumbnail_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "thumbnail_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Thumbnail Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThumbnailStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user lacks the required processing permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested processing resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProcessingErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/ProcessingErrorDetail"
              },
              {
                "$ref": "#/components/schemas/ProcessingForbiddenDetail"
              }
            ],
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ProcessingErrorResponse"
      },
...
      "ProcessingJobCreateRequest": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Idempotency Key"
          },
          "job_type": {
            "type": "string",
            "enum": [
              "vod_transcode",
              "live_transcode",
              "manifest",
              "thumbnail"
            ],
            "title": "Job Type"
          },
          "input_url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Input Url"
          },
          "output_prefix": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Output Prefix"
          },
          "renditions": {
            "items": {
              "type": "string",
              "enum": [
                "1080p",
                "720p",
                "480p"
              ]
            },
            "type": "array",
            "title": "Renditions",
            "default": [
              "1080p",
              "720p",
              "480p"
            ]
          },
          "queue": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "transcode-cpu",
                  "transcode-accelerated",
                  "manifest",
                  "thumbnail"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue"
          },
          "stream_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Id"
          },
          "recording_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recording Id"
          },
          "max_attempts": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Attempts",
            "default": 3
          },
          "timeout_seconds": {
            "type": "number",
            "maximum": 86400.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "default": 900
          },
          "workspace": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace"
          },
          "thumbnail_kinds": {
            "items": {
              "type": "string",
              "enum": [
                "poster",
                "preview",
                "timeline"
              ]
            },
            "type": "array",
            "title": "Thumbnail Kinds",
            "default": [
              "poster",
              "preview",
              "timeline"
            ]
          },
          "thumbnail_interval_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Thumbnail Interval Seconds",
            "default": 10
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key",
          "job_type",
          "input_url",
          "output_prefix"
        ],
        "title": "ProcessingJobCreateRequest"
      },
      "ProcessingJobPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProcessingJobResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ProcessingJobPage"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}